+++ crates/cargo-util/src/read2.rs
