+++ src/cargo/core/compiler/compilation.rs
@@ -4,12 +4,13 @@ use std::ffi::{OsStr, OsString};
 use std::path::PathBuf;
 
 use cargo_platform::CfgExpr;
+use cargo_util::ProcessBuilder;
 use semver::Version;
 
 use super::BuildContext;
 use crate::core::compiler::{CompileKind, Metadata, Unit};
 use crate::core::Package;
-use crate::util::{self, config, join_paths, process, CargoResult, Config, ProcessBuilder};
+use crate::util::{self, config, join_paths, CargoResult, Config};
 
 /// Structure with enough information to run `rustdoc --test`.
 pub struct Doctest {
@@ -184,7 +185,7 @@ impl<'cfg> Compilation<'cfg> {
         unit: &Unit,
         script_meta: Option<Metadata>,
     ) -> CargoResult<ProcessBuilder> {
-        let rustdoc = process(&*self.config.rustdoc()?);
+        let rustdoc = ProcessBuilder::new(&*self.config.rustdoc()?);
         let cmd = fill_rustc_tool_env(rustdoc, unit);
         let mut p = self.fill_env(cmd, &unit.pkg, script_meta, unit.kind, true)?;
         unit.target.edition().cmd_edition_arg(&mut p);
@@ -207,7 +208,13 @@ impl<'cfg> Compilation<'cfg> {
         cmd: T,
         pkg: &Package,
     ) -> CargoResult<ProcessBuilder> {
-        self.fill_env(process(cmd), pkg, None, CompileKind::Host, false)
+        self.fill_env(
+            ProcessBuilder::new(cmd),
+            pkg,
+            None,
+            CompileKind::Host,
+            false,
+        )
     }
 
     pub fn target_runner(&self, kind: CompileKind) -> Option<&(PathBuf, Vec<String>)> {
@@ -229,12 +236,12 @@ impl<'cfg> Compilation<'cfg> {
         script_meta: Option<Metadata>,
     ) -> CargoResult<ProcessBuilder> {
         let builder = if let Some((runner, args)) = self.target_runner(kind) {
-            let mut builder = process(runner);
+            let mut builder = ProcessBuilder::new(runner);
             builder.args(args);
             builder.arg(cmd);
             builder
         } else {
-            process(cmd)
+            ProcessBuilder::new(cmd)
         };
         self.fill_env(builder, pkg, script_meta, kind, false)
     }
