+++ src/cargo/util/mod.rs
@@ -4,9 +4,8 @@ pub use self::canonical_url::CanonicalUrl;
 pub use self::config::{homedir, Config, ConfigValue};
 pub use self::dependency_queue::DependencyQueue;
 pub use self::diagnostic_server::RustfixDiagnosticServer;
-pub use self::errors::{exit_status_to_string, internal, process_error, process_error_raw};
-pub use self::errors::{CargoResult, CargoResultExt, CliResult, Test};
-pub use self::errors::{CargoTestError, CliError, ProcessError};
+pub use self::errors::{internal, CargoResult, CargoResultExt, CliResult, Test};
+pub use self::errors::{CargoTestError, CliError};
 pub use self::flock::{FileLock, Filesystem};
 pub use self::graph::Graph;
 pub use self::hasher::StableHasher;
@@ -17,10 +16,8 @@ pub use self::lev_distance::{closest, closest_msg, lev_distance};
 pub use self::lockserver::{LockServer, LockServerClient, LockServerStarted};
 pub use self::paths::{bytes2path, dylib_path, join_paths, path2bytes};
 pub use self::paths::{dylib_path_envvar, normalize_path};
-pub use self::process_builder::{process, ProcessBuilder};
 pub use self::progress::{Progress, ProgressStyle};
 pub use self::queue::Queue;
-pub use self::read2::read2;
 pub use self::restricted_names::validate_package_name;
 pub use self::rustc::Rustc;
 pub use self::sha256::Sha256;
@@ -52,11 +49,9 @@ mod lockserver;
 pub mod machine_message;
 pub mod network;
 pub mod paths;
-pub mod process_builder;
 pub mod profile;
 mod progress;
 mod queue;
-mod read2;
 pub mod restricted_names;
 pub mod rustc;
 mod sha256;
